requirements.txt
streamlit
requests
google-genai
edge-tts
moviepy
python-dotenv
Pillow
